service: introduce callback for L7LB backend sync
Currently, `Service.RegisterL7LBServiceBackendSync` is built quite
Envoy specific, hence it also contains logic how to filter Service
backends by the frontendPorts that are passed to the method.

To be able to remove the dependency from the ServiceManager to Envoy
specific details (implementation and the dependency to the
Envoy xDS server), this commit refactors the backend sync registration
to receive a callback that gets called whenever a Service (and its backends)
changed.

This way, the Envoy specific details can be removed from the ServiceManager.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>