envoy: remove update-check in secret syncer
The reconciliation in the k8swatcher compared for updates by
comparing the "old" and new k8s secret. Only in cases of an actual
diff, the secret was updated in the xDS server.

`resource.Resource` no longer provides the old resource. Therefore it
would be necessary to keep track of the "old" / applied secrets to be
able to compare them. It seems as this is unnecessary, as the resource
event would only be triggered in case of a delta/update.

Therefore, this commit removes this comparison. This way we can get rid
of the mutex and secret map.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>