gateway api: no error if lb service isnt ready yet
Currently, the reconciliation of `Gateway` fails with an error if
the status of the corresponding loadbalancer service isn't ready.

Returning an error leads to an additional reconciliation and logs the error.

In case of an external loadbalancer, the status of the loadbalancer service is
never set which leads to reconciliation loops.

Therefore, with this commit, a missing status no longer results in an error. This
should also be enough in all other cases because a reconciliation should be
triggered on an update of the loadbalancer service itself.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>