endpoint: introduce EndpointCreator
Currently there are 4 ways how Endpoints are created.

* `AddHostEndpoint` when bootstrapping the daemon
* `AddIngressEndpoint` when bootstrapping the daemon
* `NewEndpointFromChangeModel` when adding or updating Endpoints via API / CNI
* `ParseEndpoint` during Endpoint restoration

All of these functions are spread over `EndpointManager` or exported functions
on the `pkg/endpoint` package directly. And all expect the Endpoint dependencies
as parameters that are passed from the daemon (tightly coupled with the daemon
struct)

In preparation to free up the Daemon from these dependencies and provide
an API that doesn't require to pass all these dependencies, this commit
introduces a new Hive Component (& Cell) `EndpointCreator`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>