fqdn/proxy: restore proxy with restored endpoints in hive lifecycle hook
This commit implements the restoration of the FQDN proxy as
hive lifecycle hook. The restored endpoints are retrieved from the
endpoint restorer via the new method `GetRestoredEndpoints`.

The `FQDNBootstrapper` interface is no longer required - therefore the bootstrap
is changed to be invoked and no longer provided. This commit doesn't further
merge the bootstrap-logic with the logic that provides the dns proxy itself.
The reason is that the new standalone fqdn proxy also depends on the proxy -
but doesn't require the lifecycle aspects (bootstrapping).

Note: This step also does not depend on any previous step within `configureDaemon`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>